requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
motor>=3.3
dnspython>=2.4
httpx>=0.27
websockets>=12.0
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
